@@ -1,34 +1,45 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
-buildscript {
-    repositories {
-        jcenter()
-    }
-    dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.0'
-        classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.4'
-        classpath 'com.github.dcendents:android-maven-gradle-plugin:1.4.1'
-    }
-}
-
-allprojects {
-    repositories {
-        jcenter()
-    }
-}
-
-ext {
-    minSdkVersion = 14
-    targetSdkVersion = 25
-    compileSdkVersion = 25
-    buildToolsVersion = '25.0.2'
-
-    // dependencies
-    supportV7 = 'com.android.support:appcompat-v7:25.3.0'
-    rxJava = 'io.reactivex.rxjava2:rxjava:2.0.2'
-    rxAndroid = 'io.reactivex.rxjava2:rxandroid:2.0.0'
-}
-
-task clean(type: Delete) {
-    delete rootProject.buildDir
+//buildscript {
+//    ext.kotlin_version = '2.1.0'
+//    repositories {
+//        google()
+//        mavenCentral()
+//    }
+//    dependencies {
+//        classpath 'com.android.tools.build:gradle:8.7.3'
+//        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+////        classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.4'
+////        classpath 'com.github.dcendents:android-maven-gradle-plugin:1.4.1'
+//    }
+//}
+//
+//allprojects {
+//    repositories {
+//        google()
+//        mavenCentral()
+//    }
+//}
+//
+//ext {
+//    minSdkVersion = 21
+//    targetSdkVersion = 34
+//    compileSdkVersion = 34
+////    buildToolsVersion = '25.0.2'
+//
+//    // dependencies
+////    supportV7 = 'androidx.appcompat:appcompat:1.7.0'
+////    rxJava = 'io.reactivex.rxjava2:rxjava:2.0.2'
+////    rxAndroid = 'io.reactivex.rxjava2:rxandroid:2.0.0'
+//}
+//
+//task clean(type: Delete) {
+//    delete rootProject.buildDir
+//}
+plugins {
+    id 'org.jetbrains.kotlin.android' version '2.1.0' apply false
+    id("com.android.application") version '8.7.3' apply false
+//    id("org.jetbrains.kotlin.android") version "1.9.23" apply false
+    id "com.google.gms.google-services" version "4.4.2" apply false
+//    id("com.google.firebase.crashlytics") version "3.0.3" apply false
 }
